Redefine GQRespValue and refactor
The old definition had several possible error representations which
weren't useful and were making trivial parts of the code look really
complicated.

We use a Vector Builder so that we can efficiently cons and concatenate
Vectors as we accumulate.

Make use of Lens a bit for modifying.

Fix modifyGQRespValue so it accepts `"data": null`

Fix gqrespValueToValue so it produces only valid graphql responses.
Rename it.