[skip ci] move all RQL mutation types to dedicated modules
This is used to break an import loop: the GraphQL context needs to
know what AST the parsers output, which is expressed via RQL types,
which in turn depends on the SchemaCache, which needs to know about
the GraphQL context...

The problem isn't new, as we already had it for select; I solved it
the same way, by creating dedicated files to hold types. This is a bit
verbose, but ultimately just moving code around.