[skip ci] move code around part 2 (mutations)
This commit finishes what was started in the previous one: move code
around to prevent an import loop. The previous commit will not compile
without this one.

They're split like this because the previous commit is entirely
unrelated to our refactor, while this one is purely related to it. If
we decide to push that change on its own to master ahead of a merge,
we will only need to cherry-pick the previous commit.