[skip ci] Fix inconsistent definitions for comparison expression types
This was a subtle issue arising from two details: our handling of
nullability and our special treatment of some Postgres types to map them
to GraphQL primitives. The former of those issues turned out to be
important to handle differently for another reason, anyway, which is now
spelled out in detail in a Note.