[skip ci] Fix required variable case in fieldWithDefault
The current logic opens the variable without marking the query
non-reusable, which can’t be right---a downstream parser might need to
open the variable, in which case we still want to mark the query
non-reusable. So we need to just invoke the downstream parser on the
value directly, leaving it still wrapped in a variable.