[skip ci] Throw fewer errors when collecting fields from fragments
Before we supported interfaces, a fragment spread could only have been valid if
its type condition was equal to the name of the object whose selection set we're
parsing.  However, now that we are supporting interfaces, this is getting much
more involved.  So this disables any error checking until we figure out how to
do it.