[skip ci] Get some remote schemas tests to pass
I am updating one of the integration tests: sure, `Hello` is a type exposed by
the remote schema. But it is the query root, which has no justification for
being exposed through our introspection. It would be a mistake to expose the
query root of all remote schemas: because there may be several remote schemas
whose query root is called `query_root` (such as when they are Hasura
instances).  So we shouldn't expect to find that name in the types.