[skip ci] fixed empty updates check
This one is tricky: the code used to check that at least one column
was modified, meaning at least one operation in the arguments of the
update, or at least one preset column. However, this is not the same
thing as what the old code used to do: the old code allows for an
empty mutation, as long as at least one operator is present in the
request, even if empty; the following is therefore valid:

    mutation foo {
      update_author(
        where: {id: {_eq: 1}},
        _set: {}
      ) {
        affected_rows
      }
    }

The old code would then check the update object when generating the
SQL transaction, and just return an empty mutation result, which is
something our new code didn't do.

This commit reverts to master's behaviour, by changing the validity
check and adding the new SQL generation case for empty mutations.