[skip ci] disable check that all variables are used
While this is mandated by the spec, we currently do not perform that
check on master, which means that introducing it now might cause
regressions for users that (perhaps unknowingly) rely on this
oversight.

We will probably want to enable it at a later point, after we're done
merging in master, and with the ability to control it with a flag,
perhaps alongside a major release if we want to make this new (and
correct) behaviour the default.