Consistently log the request_id at the same level
Before this change, the server is logging requests with a different
shape.

- HTTPLog

```json
{
  "type": "http-log",
  "timestamp": "2020-11-23T17:30:41.782+0100",
  "level": "info",
  "detail": {
    "operation": {
      "query_execution_time": 8.6201e-05,
      "user_vars": {
        "x-hasura-role": "admin"
      },
      "request_id": "f579069a-23fc-4119-b2d8-4bab1a4e29e6",
      "response_size": 347,
      "request_read_time": 3.884e-06
    },
    "http_info": {
      "status": 200,
      "http_version": "HTTP/1.1",
      "url": "/foo",
      "ip": "127.0.0.1",
      "method": "GET",
      "content_encoding": "gzip"
    }
  }
}
```

- QueryLog

```json
{
"detail":  {
  "http_info" {...}
  "operation" {
    "query_execution_time": 0.001201056,
    "request_id": "e9bd9dd2-f5b3-4e26-9d88-9a39cff5acbb"
    ...
  }
},
"level": "info",
"timestamp": "2020-11-23T16:48:07.173+0000",
"type":	"http-log"
}
```

Notice that the `request_id` is nested in different levels in both
examples. This makes correlating the requests a bit more difficult than
it should be.

For instance, services like DataDog allow you to extract a json path as
a facet that can be used in searches. Since the `request_id` is at a
different level, it is not possible to correlate using the `request_id`.

After this change, the `QueryLog` will look like

```json
{
"detail":  {
  "http_info" {...}
  "operation" {
    "query_execution_time": 0.001201056,
    "request_id": "e9bd9dd2-f5b3-4e26-9d88-9a39cff5acbb"
    ...
  },
  "request_id": "e9bd9dd2-f5b3-4e26-9d88-9a39cff5acbb"
},
"level": "info",
"timestamp": "2020-11-23T16:48:07.173+0000",
"type":	"http-log"
}
```

The `request_id` is still present in the `operation` key for backwards
compatibility.