Adding seeds documentation config-reference.rst
Document hasn't been updated with references to seeds. Apologies for the formatting of adding a new row to the table. Shows a bunch of changes but really I just added one row. I felt it was more important to group metadata/migrations/seeds together as they all follow the same pattern, instead of adding it at the bottom for a cleaner commit.