ci: run stylish haskell on all incoming PRs
### Description

This PR adds two new github actions: one in the OSS repo, one in the monorepo. The OSS action runs `stylish-haskell` on all files touched by the PR, and displays a warning for each file that was in need of a formatting. The monorepo action does the same thing, with a twist: if the branch is not a shadow copy of an OSS branch, we assume that it is a local branch, and simply push a new commit with the changes.

Furthermore, this PR upgrades our stylish-haskell config to add record formatting, as close as possible to our styleguide.

Both actions use the standard stylish-haskell, not our modified fork.

### Known limitation

The monorepo action does not handle forks: pushing to the branch will fail, and checking the branch out might fail too. This is probably acceptable since we don't use forks with the monorepo, but it wouldn't be hard to handle that gracefully.

GitOrigin-RevId: 814138c5b9826098e2e4ea192778fc0d93fbe390