Clarify many-to-many foreign key
Clarify that foreign keys are meant to be added in the join table, not on each table to be joined. If you attempt to add foreign keys to the individual tables you get an error that is difficult to parse.
Also in some cases, a join table could join two objects more than once and therefore using a primary key might be best.